Make the default cache time more explicit in code
To avoid confusion, make the cache time and cache toggle for a
given subsystem configurable via the more friendly
`get_memoization_decorator`. This has the added benefit of
making the resulting decorated methods easier to read as they
are only decorated with the resulting function not needing
the decorator with the same arguments repeatedly.

This change also creates the `get_expiration_time_fn` function
that better describes the fallthrough behavior of the [cache]
expiration_time config value.

This patch updates the revocation event cache time to be in the
proper [revoke] section of the config for consistency with the
other sections.

DocImpact: revocation_cache_time moved from [token] config
section to [revoke]/cache_time

Co-Authored-By: David Stanek <dstanek@dstanek.com>
Change-Id: I5c4e5f509ab82b060938591655e6c615ac7d38dd