Move optional backend dependencies into setup.cfg [extras]
The dogpile and mongo backends are optional so therefore the
dependencies that are only used for those backends are optional.

This makes use of the 'extras' capability in setuptools to declare those
as optional dependencies if you're using those backends.

If a consumer of oslo.cache wants to use the dogpile backend, for
example, then they install oslo.cache like 'sudo pip install
oslo.cache[dogpile]'.

Note that the dogpile dependency is left in requirements.txt because
the oslo_cache.core module has an unconditional import on that library.

Change-Id: I28332964d568bda405fb8d1376d50db49d288b94