Correct help text for backend option
The help for the backend option said that the memcache_pool backend
is recommended.

memcache_pool is there to keep a limited number of thread-associated
connections open rather than a connection for every thread. If you
don't have a huge number of threads it doesn't offer anything.
Keystone is an example of a service where memcache_pool doesn't
improve things -- eventlet isn't supported anymore and more threads
is not useful due to GIL.

As such, the help text is corrected to say to use memcache_pool for
eventlet and dogpile.cache.memcached for servers like keystone.

The help text also mentioned devstack. This is removed because
devstack is a developer tool so shouldn't be mentioned in text
targeted to operators.

Also, fixed the preceeding comment to remove redundant information
and to not mention keystone since this code isn't in keystone
anymore.

Change-Id: I1ffbf23feac623fe0bfbbe49fb3e0159dedf4cfe