Config dict arguments.url must be a list
For memcache, arguments.url is assumed to be a list, this means
that the underlying memcache library could not handle a case where
a string (comma separated) was passed in. The idea is to force a
split on ',' in the special case of backend_argument=url:...
so that a list is passed to the backend.

Change-Id: I65afb46455423d7f96e17c471330c8d74dc17cac
Closes-Bug: #1743036
(cherry picked from commit 685a05edfef24cb2300df24aa3417ad0ecd2c2c2)