Pass `flush_on_reconnect` to memcache pooled backend
If a memcache server disappears and then reconnects when multiple memcache
servers are used (specific to the python-memcached based backends) it is
possible that the server will contain stale data. The default is now to
supply the ``flush_on_reconnect`` optional argument to the backend. This
means that when the service connects to a memcache server, it will flush
all cached data in the server. The pooled backend is more likely to
run into issues with this as it does not explicitly use a thread.local
for the client. The non-pooled backend was not touched, it is not
the recommended production use-case.

See the help from python-memcached:

    @param flush_on_reconnect: optional flag which prevents a
	scenario that can cause stale data to be read: If there's more
	than one memcached server and the connection to one is
	interrupted, keys that mapped to that server will get
	reassigned to another. If the first server comes back, those
	keys will map to it again. If it still has its data, get()s
	can read stale data that was overwritten on another
	server. This flag is off by default for backwards
	compatibility.

Change-Id: I3e335261f749ad065e8abe972f4ac476d334e6b3
closes-bug: #1819957