[doc] tell to prefix ipv6 hosts with inet6
Some environments with ipv6 used to connect with memcached servers
faced issues by omitting to prefix the host with the address family to
use.

Indeed libraries like python-memcached waiting for a prefix to
initialize the socket with the right address family. If the prefix
isn't given and that the host refer or bind an ipv6 then the socket
is initialized with an address family which correspond to an IPv4
and the connection doesn't work properly.

Openstack's puppet libraries already fixed that by introducing new
feature to automatically prefix the address if needed [1][2].

Anyway it could be worth to add more documentation on oslo.cache
to give more information to customers and avoid similar issues. These
changes aim to provide this documentation.

[1] https://review.opendev.org/#/q/status:merged+branch:master+topic:memcached-inet6-prefix
[2] https://review.opendev.org/#/c/661136/

Change-Id: I5e0b4818300348efe4625500792d9161281ed45a