Switch to newer openstackdocstheme and reno versions
Switch to openstackdocstheme 2.2.0 and reno 3.1.0 versions. Using
these versions will allow especially:
* Linking from HTML to PDF document
* Allow parallel building of documents
* Fix some rendering problems

Update Sphinx version as well.

openstackdocstheme renames some variables, so follow the renames. A
couple of variables are also not needed anymore, remove them.

Set openstackdocs_auto_name to use project as name.

Fix wrong indent in index.rst that lead to a grey bar on the left of the
toc.

Depends-On: https://review.opendev.org/728938
Change-Id: If82be6fbfb12cedd8f9dcfb4468415e085d1c821