Add retry mechanisms to oslo.cache
This patch specifies a set of options required to setup a retry
context.  The context built from those options can later on be
passed to any of the oslo.cache backends that supports pymemcache's
retry mechanisms.

This patch also sets up the retry mechanisms context based on
the configuration option passed via oslo.config and adds it
as an argument to be passed to the selected oslo.cache backend.

This patch is needed to fix a TLS issue on stable branches introduced by
pymemcache (since train), where if a cluster node disappear the client
will fail without retrying to reconnect or to switch to an other node of
the cluster.

Partial-Bug: #1959562
Change-Id: I6c1a4872d7cf19e3a55c676ef4b4200f18e08f2c
(cherry picked from commit 42bf82d5505a0def3d22972d67fde836f5091ed1)