Add retry mechanisms to oslo.cache
This patch specifies a set of options required to setup a retry
context.  The context built from those options can later on be
passed to any of the oslo.cache backends that supports pymemcache's
retry mechanisms.

This patch also sets up the retry mechanisms context based on
the configuration option passed via oslo.config and adds it
as an argument to be passed to the selected oslo.cache backend.

Change-Id: I6c1a4872d7cf19e3a55c676ef4b4200f18e08f2c