Do not mark hosts as alive when they are all dead
This feature is supposed to check for alive servers more frequently when they are all down.
Unfortunately, it does more harm than good.

I have made several tests with keystone and here is what I saw when all memcached backends were unreachable:
1. memcached services were stopped but their hosts were alive(keystone received 'connection refused') - everything was working fine
2. memcached backends and their hosts were completely unreachable(keystone received 'no route to host') - the whole keystone stopped responding

Keystone should be working fine even if all memcached backends are down. In this case it wasn't.
After I deleted the discussed functionality from the code, keystone was finally able to handle failure of all memcached backends.
I believe this issue is not only limited to keystone, I just used it as an example.

Change-Id: Ib3e3028d967c166d21b60cf4cb7c9d5dc82a8fe7