Expose pymemcache's HashClient public params
This patch expose a couple of pymemcache's HashClient public
params that can be useful to configure HA and failover for
clustered memcached servers.

These options can be used in addition of the previously added
retrying mechanismes.

This patch rely on recent changes [1] of dogpile.cache that
aim to expose these options too.

[1] https://gerrit.sqlalchemy.org/c/sqlalchemy/dogpile.cache/+/3528

Change-Id: I24fc853db4237c08b87871ddff1b3ced27cc7fce