[stable-only] Fix skipIf conditional of pymemcache features
Those conditionals are reversed.
Some pymemcache features are not supported by the version of pymemcache
supported in upper-constraints.

For more context, we needed to backport the socket keepalive feature and
the retrying feature to fix a functional issue introduced by pymemcache
when we introduced the support of TLS months ago. However to add the
TLS support people backported patches that was not synced with the
requirements versions provided by upper-constraints, and so that mean
that they bypassed a global requirements on stable branch. So we are
stuck with this approach with the backport of the TLS fix, as upgrading
requirements on stable branches is forbiden by our policy. This is why
some tests are skipped.

Change-Id: Ia248e09942a858a02a8bf64ca4869f8813348d94
Co-Authored-By: Hervé Beraud <hberaud@redhat.com>