Fix description of memcache_servers
Current description says the option is used by only two backends but
in fact this option is used by all backends dependent on memcached.

Note that the inet6 prefix is specific to the python-memcached library
and should not be used if a backend dependent on different libraries
(eg. pymemcache) is used. This change describes that point, too.

Change-Id: I3c967841c94b9409096c9b9591f9fb162db7a1ac