Clean-up memcache connection sockets
The ConnectionPool currently opens sockets, but never closes them. As a
result, any client using oslo.cache memcache pool leaves sockets in a
CLOSE_WAIT state, with the source port not being re-usable.

In our production system, were we have A LOT of activity, this creates
a storm of non-reusable ports: all source ports are in use, making the
node unuseable for other things.

This patch adds a __del__ destructor closing inactive connections,
fixing the issue.

Closes-Bug: #2043121
Change-Id: I09d632346c76d1aff7c534f0d040162d1985f548