Fail if tls_enabled is True but backend does not support it
The tls_enabled option works properly only for the following backends.
 - oslo_cache.memcache_pool
 - dogpile.cache.bmemcache
 - dogpile.cache.pymemcache

Currently the tls options are all ignored in case a different backend
is used, but this does not allow users to notice that TLS is not
enabled contrary to their expectations.

This introduces the validation to ensure an appropriate backend is used
when tls is enabled.

NOTE:
The oslo_cache.memcache_pool bakcend supports tls_enabled only when
sasl_enabled is true, which can be fixed separately.

Change-Id: Ib967bf8cb21fb97fff94a6d6cb5983374e4798eb