Run mypy from tox
This avoids the need to duplicate our dependency list in multiple
places and allows us to take advantage of tox's dependency management
infrastructure, to ensure we always get the latest and greatest version
of a package allowed by upper-constraints.

While here, we also update versions of other pre-commit hooks and
simplify the mypy configuration, aligning it with that of other typed
oslo projects.

Change-Id: I9ca3ec0d6f2de19f52f3d4c46fcbc43f9ed3a5b4
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>